Extract BasicMessageView from ChatMessageView to fix stack overflow crash on physical devices
